Allow config files to contain a root element named 'gpg'.
Before this change if the config file had a 'gpg' root element it would
override the figgypy gpg object. I'm unsure if this really broke
anything but it was a bit awkward that an internal only object existed
in the config object (as conf.gpg) when the file didn't have a gpg
element. I was also awkward that a config file could change the internal
workings of the config object.