Add parameter for gpg_config and change instantiation
The reason for this is that new versions of gnupg cause the core library
to choke and fail. So now we are using the forked version which is able
to continue to work. This also improves the ability to override defaults
with environment or passing configuration into the instance directly.